+++ .gitignore
@@ -31,3 +31,7 @@ build/
 
 ### VS Code ###
 .vscode/
+
+### OTHER ###
+.env
+.gitignore
