--- compose.yaml
@@ -1,31 +1,0 @@
-services:
-  postgres:
-    container_name: postgres-sql-clothing
-    image: postgres
-    environment:
-      POSTGRES_USER: "username"
-      POSTGRES_PASSWORD: "password"
-      PGDATA: /var/lib/postgresql/data
-      POSTGRES_DB: clothing_postgres
-    volumes:
-      - postgres:/var/lib/postgresql/data
-    ports:
-      - "5432:5432"
-    networks:
-      - clothing-dev
-    restart: unless-stopped
-  mail-dev:
-    container_name: mail-dev-clothing
-    image: maildev/maildev
-    ports:
-      - "1080:1080"
-      - "1025:1025"
-
-
-networks:
-  clothing-dev:
-    driver: bridge
-
-volumes:
-  postgres:
-    driver: local
