--- src/main/java/org/tn/subscriptiontool/core/auth/config/CorsConfig.java
@@ -1,27 +1,0 @@
-package org.tn.subscriptiontool.core.auth.config;
-
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.web.servlet.config.annotation.CorsRegistry;
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-
-@Configuration
-public class CorsConfig implements WebMvcConfigurer {
-    @Value("${application.host.domain}")
-    private String domain;
-
-    @Value("${application.host.frontend.port}")
-    private String frontEndPort;
-
-    @Override
-    public void addCorsMappings(CorsRegistry registry) {
-        registry.addMapping("/**")
-                .allowedOrigins(
-                        "http://127.0.0.1:" + frontEndPort,
-                        "http://localhost:" + frontEndPort,
-                        domain + ":" + frontEndPort
-                ).allowedMethods("GET", "POST", "PUT", "DELETE")
-                .allowedHeaders("*")
-                .allowCredentials(true);
-    }
-}
