--- src/main/java/org/tn/subscriptiontool/core/auth/config/SecurityConfig.java
@@ -1,57 +1,0 @@
-package org.tn.subscriptiontool.core.auth.config;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.security.authentication.AuthenticationProvider;
-import org.springframework.security.config.annotation.method.configuration.EnableMethodSecurity;
-import org.springframework.security.config.annotation.web.builders.HttpSecurity;
-import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
-import org.springframework.security.config.annotation.web.configurers.AbstractHttpConfigurer;
-import org.springframework.security.web.SecurityFilterChain;
-import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
-import org.tn.subscriptiontool.core.auth.authorization.providers.EmailTokenAuthenticationProvider;
-import org.tn.subscriptiontool.core.auth.filters.JwtFilter;
-
-import static org.springframework.security.config.Customizer.withDefaults;
-import static org.springframework.security.config.http.SessionCreationPolicy.STATELESS;
-
-
-@Configuration
-@EnableWebSecurity
-@RequiredArgsConstructor
-@EnableMethodSecurity(securedEnabled = true)
-public class SecurityConfig {
-    private final AuthenticationProvider daoAuthenticationProvider;
-    private final EmailTokenAuthenticationProvider emailTokenAuthenticationProvider;
-    private final JwtFilter jwtFilter;
-
-    @Bean
-    public SecurityFilterChain securityFilterChain(HttpSecurity http) throws Exception {
-        return http
-                .cors(withDefaults())
-                .csrf(AbstractHttpConfigurer::disable)
-                .authenticationProvider(daoAuthenticationProvider)
-                .authenticationProvider(emailTokenAuthenticationProvider)
-                .authorizeHttpRequests(request -> {
-                    request.requestMatchers(
-                            "/account/register",
-                            "/account/login",
-                            "/account/activate",
-                            "/v3/api-docs",
-                            "/v3/api-docs/**",
-                            "/swagger-resources",
-                            "/configuration/ui",
-                            "/configuration/security",
-                            "/swagger-ui/**",
-                            "/webjars/**",
-                            "/swagger-ui.html"
-                    ).permitAll()
-                            .anyRequest()
-                            .authenticated();
-                })
-                .sessionManagement(session -> session.sessionCreationPolicy(STATELESS))
-                .addFilterBefore(jwtFilter, UsernamePasswordAuthenticationFilter.class)
-                .build();
-    }
-}
