--- src/main/java/org/tn/subscriptiontool/core/auth/models/EmailTemplateName.java
@@ -1,14 +1,0 @@
-package org.tn.subscriptiontool.core.auth.models;
-
-import lombok.Getter;
-
-@Getter
-public enum EmailTemplateName {
-    ACTIVATE_ACCOUNT("activate_account");
-
-    private final String name;
-
-    EmailTemplateName(String name) {
-        this.name = name;
-    }
-}
