--- src/main/java/org/tn/subscriptiontool/core/auth/models/Roles.java
@@ -1,38 +1,0 @@
-package org.tn.subscriptiontool.core.auth.models;
-
-import com.fasterxml.jackson.annotation.JsonIgnore;
-import jakarta.persistence.*;
-import lombok.*;
-import org.springframework.data.annotation.CreatedDate;
-import org.springframework.data.jpa.domain.support.AuditingEntityListener;
-
-import java.time.LocalDateTime;
-import java.util.List;
-
-@Getter
-@Setter
-@Builder
-@AllArgsConstructor
-@NoArgsConstructor
-@Entity
-@EntityListeners(AuditingEntityListener.class)
-public class Roles {
-    @Id
-    @GeneratedValue(strategy = GenerationType.SEQUENCE)
-    private Long ID;
-
-    @Column(unique = true)
-    private String name;
-
-    @ManyToMany(mappedBy = "roles")
-    @JsonIgnore
-    private List<User> users;
-
-    @CreatedDate
-    @Column(nullable = false, updatable = false)
-    private LocalDateTime createdDate;
-
-    @CreatedDate
-    @Column(insertable = false)
-    private LocalDateTime lastModifiedDate;
-}
