--- src/main/java/org/tn/subscriptiontool/core/auth/models/Token.java
@@ -1,29 +1,0 @@
-package org.tn.subscriptiontool.core.auth.models;
-
-import jakarta.persistence.*;
-import jakarta.transaction.Transactional;
-import lombok.*;
-
-import java.time.LocalDateTime;
-
-@Getter
-@Setter
-@Builder
-@AllArgsConstructor
-@NoArgsConstructor
-@Entity
-@Transactional
-public class Token {
-    @Id
-    @GeneratedValue(strategy = GenerationType.SEQUENCE)
-    private Long ID;
-    private String token;
-
-    private LocalDateTime createdAt;
-    private LocalDateTime expiresAt;
-    private LocalDateTime validatedAt;
-
-    @ManyToOne
-    @JoinColumn(name = "user_id", nullable = false)
-    private User user;
-}
