--- src/main/java/org/tn/subscriptiontool/core/auth/repository/RoleRepository.java
@@ -1,9 +1,0 @@
-package org.tn.subscriptiontool.core.auth.repository;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.tn.subscriptiontool.core.auth.models.Roles;
-
-import java.util.Optional;
-
-public interface RoleRepository extends JpaRepository<Roles, Long> {
-    Optional<Roles> findByName(String name);
-}
