--- src/main/java/org/tn/subscriptiontool/core/auth/services/EmailService.java
@@ -1,68 +1,0 @@
-package org.tn.subscriptiontool.core.auth.services;
-
-import jakarta.mail.MessagingException;
-import jakarta.mail.internet.MimeMessage;
-import lombok.RequiredArgsConstructor;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.mail.javamail.JavaMailSender;
-import org.springframework.mail.javamail.MimeMessageHelper;
-import org.springframework.scheduling.annotation.Async;
-import org.springframework.stereotype.Service;
-import org.thymeleaf.context.Context;
-import org.thymeleaf.spring6.SpringTemplateEngine;
-import org.tn.subscriptiontool.core.auth.models.EmailTemplateName;
-
-import java.util.HashMap;
-import java.util.Map;
-
-import static java.nio.charset.StandardCharsets.UTF_8;
-import static org.springframework.mail.javamail.MimeMessageHelper.MULTIPART_MODE_MIXED;
-
-@Service
-@RequiredArgsConstructor
-public class EmailService {
-    private final JavaMailSender mailSender;
-    private final SpringTemplateEngine templateEngine;
-    @Value("${application.security.mailing.from}")
-    private String from;
-
-    @Async
-    public void sendEmail(
-            String to,
-            String username,
-            EmailTemplateName emailTemplateName,
-            String activationCode,
-            String subject
-    ) throws MessagingException {
-        String templateName;
-        if (emailTemplateName == null) {
-            templateName = "activate_account";
-        } else {
-            templateName = emailTemplateName.name();
-        }
-
-        MimeMessage mimeMessage = mailSender.createMimeMessage();
-        MimeMessageHelper helper = new MimeMessageHelper(
-                mimeMessage,
-                MULTIPART_MODE_MIXED,
-                UTF_8.name()
-        );
-
-        Map<String, Object> properties = new HashMap<>();
-        properties.put("username", username);
-        properties.put("activationCode", activationCode);
-
-        Context context = new Context();
-        context.setVariables(properties);
-
-        helper.setFrom(from);
-        helper.setTo(to);
-        helper.setSubject(subject);
-
-        String template = templateEngine.process(templateName, context);
-
-        helper.setText(template, true);
-
-        mailSender.send(mimeMessage);
-    }
-}
