--- src/main/java/org/tn/subscriptiontool/core/logger/LoggerService.java
@@ -1,22 +1,0 @@
-package org.tn.subscriptiontool.core.logger;
-
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import org.springframework.stereotype.Component;
-
-@Component
-public class LoggerService {
-    private final Logger logger = LoggerFactory.getLogger(LoggerService.class);
-
-    public void logInfo(String message) {
-        logger.info(message);
-    }
-
-    public void logWarning(String message) {
-        logger.warn(message);
-    }
-
-    public void logError(String message) {
-        logger.error(message);
-    }
-}
