+++ src/main/java/org/tn/subscriptiontool/core/security/exceptions/GlobalExceptionHandler.java
@@ -1,0 +1,104 @@
+package org.tn.subscriptiontool.core.security.exceptions;
+
+import jakarta.mail.MessagingException;
+import org.springframework.http.ResponseEntity;
+import org.springframework.security.authentication.BadCredentialsException;
+import org.springframework.security.authentication.DisabledException;
+import org.springframework.security.authentication.LockedException;
+import org.springframework.web.bind.MethodArgumentNotValidException;
+import org.springframework.web.bind.annotation.ExceptionHandler;
+import org.springframework.web.bind.annotation.RestControllerAdvice;
+import org.tn.subscriptiontool.core.security.payloads.responses.ExceptionResponse;
+
+import java.util.HashSet;
+import java.util.Set;
+
+import static org.springframework.http.HttpStatus.*;
+import static org.tn.subscriptiontool.core.security.exceptions.ErrorCodes.*;
+
+@RestControllerAdvice
+public class GlobalExceptionHandler {
+
+    @ExceptionHandler(LockedException.class)
+    public ResponseEntity<ExceptionResponse> handleException(LockedException exception) {
+        return ResponseEntity
+                .status(UNAUTHORIZED)
+                .body(
+                        ExceptionResponse.builder()
+                                .businessExceptionDescription(ACCOUNT_LOCKED.getDescription())
+                                .error(!exception.getMessage().isBlank()
+                                        ? exception.getMessage() :
+                                        ACCOUNT_LOCKED.getDescription()
+                                )
+                                .build()
+                );
+    }
+
+    @ExceptionHandler(DisabledException.class)
+    public ResponseEntity<ExceptionResponse> handleException(DisabledException exception) {
+        return ResponseEntity
+                .status(UNAUTHORIZED)
+                .body(
+                        ExceptionResponse.builder()
+                                .businessExceptionDescription(ACCOUNT_DISABLED.getDescription())
+                                .error(exception.getMessage())
+                                .build()
+                );
+    }
+
+    @ExceptionHandler(BadCredentialsException.class)
+    public ResponseEntity<ExceptionResponse> handleException(BadCredentialsException exception) {
+        return ResponseEntity
+                .status(UNAUTHORIZED)
+                .body(
+                        ExceptionResponse.builder()
+                                .businessExceptionDescription(BAD_CREDENTIALS.getDescription())
+                                .error(!exception.getMessage().isBlank()
+                                        ? exception.getMessage() :
+                                        BAD_CREDENTIALS.getDescription()
+                                ).build()
+                );
+    }
+
+    @ExceptionHandler(MessagingException.class)
+    public ResponseEntity<ExceptionResponse> handleException(MessagingException exception) {
+        return ResponseEntity
+                .status(INTERNAL_SERVER_ERROR)
+                .body(
+                        ExceptionResponse.builder()
+                                .error(exception.getMessage())
+                                .build()
+                );
+    }
+
+    @ExceptionHandler(MethodArgumentNotValidException.class)
+    public ResponseEntity<ExceptionResponse> handleException(MethodArgumentNotValidException exception) {
+        Set<String> errors = new HashSet<>();
+
+        exception.getBindingResult().getAllErrors().forEach(error -> {
+            var errorMessage = error.getDefaultMessage();
+            errors.add(errorMessage);
+        });
+
+        return ResponseEntity
+                .status(BAD_REQUEST)
+                .body(
+                        ExceptionResponse.builder()
+                                .validationErrors(errors)
+                                .build()
+                );
+    }
+
+    @ExceptionHandler(Exception.class)
+    public ResponseEntity<ExceptionResponse> handleException(Exception exception) {
+        exception.printStackTrace();
+        return ResponseEntity
+                .status(INTERNAL_SERVER_ERROR)
+                .body(
+                        ExceptionResponse.builder()
+                                .businessExceptionDescription("Internal error, contact the admin.")
+                                .error(exception.getMessage())
+                                .build()
+                );
+    }
+}
