+++ src/main/java/org/tn/subscriptiontool/core/security/models/EmailTemplateName.java
@@ -1,0 +1,14 @@
+package org.tn.subscriptiontool.core.security.models;
+
+import lombok.Getter;
+
+@Getter
+public enum EmailTemplateName {
+    ACTIVATE_ACCOUNT("activate_account");
+
+    private final String name;
+
+    EmailTemplateName(String name) {
+        this.name = name;
+    }
+}
