+++ src/main/java/org/tn/subscriptiontool/core/security/payloads/responses/ExceptionResponse.java
@@ -1,0 +1,22 @@
+package org.tn.subscriptiontool.core.security.payloads.responses;
+
+import com.fasterxml.jackson.annotation.JsonInclude;
+import lombok.*;
+
+import java.util.Map;
+import java.util.Set;
+
+@Getter
+@Setter
+@Builder
+@AllArgsConstructor
+@NoArgsConstructor
+@JsonInclude(JsonInclude.Include.NON_EMPTY)
+public class ExceptionResponse {
+
+    private Integer STATUS;
+    private String businessExceptionDescription;
+    private String error;
+    private Set<String> validationErrors;
+    private Map<String, String> errors;
+}
