+++ src/main/java/org/tn/subscriptiontool/core/security/services/JwtService.java
@@ -1,0 +1,165 @@
+package org.tn.subscriptiontool.core.security.services;
+
+import io.jsonwebtoken.Claims;
+import io.jsonwebtoken.Jwts;
+import io.jsonwebtoken.SignatureAlgorithm;
+import io.jsonwebtoken.io.Decoders;
+import io.jsonwebtoken.security.Keys;
+import lombok.RequiredArgsConstructor;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.security.core.GrantedAuthority;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.stereotype.Service;
+
+import javax.crypto.SecretKey;
+import java.security.Key;
+import java.util.Date;
+import java.util.HashMap;
+import java.util.Map;
+import java.util.function.Function;
+
+/**
+ * The {@code JwtService} class provides functionality for generating, validating,
+ * and extracting information from JSON Web Tokens (JWTs).
+ *
+ * <p>This service handles the creation of JWTs, claims extraction, and
+ * token validation, as well as managing expiration settings for the tokens.</p>
+ */
+@Service
+@RequiredArgsConstructor
+public class JwtService {
+    @Value("${application.security.jwt.secret-key}")
+    private String secretKey;
+
+    @Value("${application.security.jwt.expiration}")
+    private long jwtExpiration;
+
+    /**
+     * Extracts the username from the given JWT token.
+     *
+     * @param token the JWT token from which the username will be extracted.
+     * @return the username contained in the JWT token.
+     */
+    public String extractUsername(String token) {
+        return extractClaim(token, Claims::getSubject);
+    }
+
+    /**
+     * Extracts a specific claim from the JWT token using the provided claim resolver.
+     *
+     * @param token the JWT token from which to extract the claim.
+     * @param claimResolver a function to resolve the claim from the claims object.
+     * @param <T> the type of the claim to extract.
+     * @return the extracted claim value.
+     */
+    public <T> T extractClaim(String token, Function<Claims, T> claimResolver) {
+        final Claims claims = extractAllClaims(token);
+        return claimResolver.apply(claims);
+    }
+
+    /**
+     * Extracts all claims from the given JWT token.
+     *
+     * @param token the JWT token from which to extract all claims.
+     * @return the claims contained in the JWT token.
+     */
+    private Claims extractAllClaims(String token) {
+        return Jwts
+                .parser()
+                .verifyWith(getSignInKey())
+                .build()
+                .parseSignedClaims(token)
+                .getPayload();
+    }
+
+    /**
+     * Generates a JWT token for the specified user details.
+     *
+     * @param userDetails the user details for which the token will be generated.
+     * @return the generated JWT token.
+     */
+    public String generateToken(UserDetails userDetails) {
+        return generateToken(new HashMap<>(), userDetails);
+    }
+
+    /**
+     * Generates a JWT token with specific claims for the specified user details.
+     *
+     * @param claims additional claims to include in the token.
+     * @param userDetails the user details for which the token will be generated.
+     * @return the generated JWT token.
+     */
+    public String generateToken(Map<String, Object> claims, UserDetails userDetails) {
+        return buildToken(claims, userDetails, jwtExpiration);
+    }
+
+    /**
+     * Validates the given JWT token against the user details.
+     *
+     * @param token the JWT token to validate.
+     * @param userDetails the user details to validate against.
+     * @return {@code true} if the token is valid, {@code false} otherwise.
+     */
+    public boolean isTokenValid(String token, UserDetails userDetails) {
+        final String userEmail = extractUsername(token);
+        return (userEmail.equals(userDetails.getUsername()) && !isTokenExpired(token));
+    }
+
+    /**
+     * Checks if the given JWT token is expired.
+     *
+     * @param token the JWT token to check.
+     * @return {@code true} if the token is expired, {@code false} otherwise.
+     */
+    private boolean isTokenExpired(String token) {
+        return extractExpiration(token).before(new Date());
+    }
+
+
+    /**
+     * Extracts the expiration date from the given JWT token.
+     *
+     * @param token the JWT token from which to extract the expiration date.
+     * @return the expiration date of the token.
+     */
+    private Date extractExpiration(String token) {
+        return extractClaim(token, Claims::getExpiration);
+    }
+
+    /**
+     * Builds a JWT token with the specified claims, user details, and expiration time.
+     *
+     * @param claims additional claims to include in the token.
+     * @param userDetails the user details for which the token will be generated.
+     * @param jwtExpiration the expiration time for the token.
+     * @return the constructed JWT token.
+     */
+    private String buildToken(
+            Map<String, Object> claims,
+            UserDetails userDetails,
+            long jwtExpiration
+    ){
+        var authorities = userDetails.getAuthorities()
+                .stream()
+                .map(GrantedAuthority::getAuthority)
+                .toList();
+        return Jwts.builder()
+                .claims(claims)
+                .subject(userDetails.getUsername())
+                .issuedAt(new Date(System.currentTimeMillis()))
+                .expiration(new Date(System.currentTimeMillis() + jwtExpiration))
+                .claim("authorities", authorities)
+                .signWith(getSignInKey(), SignatureAlgorithm.HS384)
+                .compact();
+    }
+
+    /**
+     * Retrieves the signing key used for signing the JWTs.
+     *
+     * @return the signing key as a {@link Key} object.
+     */
+    private SecretKey getSignInKey() {
+        byte[] keyBytes = Decoders.BASE64.decode(secretKey);
+        return Keys.hmacShaKeyFor(keyBytes);
+    }
+}
